drivers: sensor: scd4x: fix overflow in temperature offset getter
scd4x_get_temperature_offset() computed the fractional part with a
32-bit accumulator, so (temp % 0xFFFF) * 1000000 overflowed for most
raw values and reported a corrupted offset. Widen the intermediate to
int64_t, matching the conversion in scd4x_channel_get().

+++ drivers/sensor/sensirion/scd4x/scd4x.c
@@ -311,12 +311,12 @@ static int scd4x_get_temperature_offset(const struct device *dev, struct sensor_
 		return ret;
 	}
 
-	int32_t temp;
+	int64_t temp;
 
 	/*Calculation from Datasheet*/
 	temp = sys_get_be16(rx_buf) * SCD4X_MAX_TEMP;
 	val->val1 = (int32_t)(temp / 0xFFFF);
-	val->val2 = ((temp % 0xFFFF) * 1000000) / 0xFFFF;
+	val->val2 = (int32_t)(((temp % 0xFFFF) * 1000000) / 0xFFFF);
 
 	return 0;
 }
